requirements.txt
altgraph==0.17.4
asgiref==3.8.1
Django==4.2.16
image==1.5.33
importlib_metadata==8.5.0
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.21.0
